+++ src/SqlClient/AssemblyInfo.fs
@@ -4,9 +4,9 @@ open System.Reflection
 [<assembly: AssemblyTitleAttribute("SqlClient")>]
 [<assembly: AssemblyProductAttribute("FSharp.Data.SqlClient")>]
 [<assembly: AssemblyDescriptionAttribute("SqlClient F# type providers")>]
-[<assembly: AssemblyVersionAttribute("1.2.24")>]
-[<assembly: AssemblyFileVersionAttribute("1.2.24")>]
+[<assembly: AssemblyVersionAttribute("1.2.26")>]
+[<assembly: AssemblyFileVersionAttribute("1.2.26")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] Version = "1.2.24"
+    let [<Literal>] Version = "1.2.26"